requirements.txt
# requirements.txt

# FastAPI 框架
fastapi==0.111.0

# ASGI 服务器
uvicorn[standard]==0.29.0

# 数据库ORM
SQLAlchemy==2.0.30

# 数据库迁移工具
alembic==1.13.1

# Pydantic (V2版本), 用于数据验证和设置管理
pydantic==2.7.4
pydantic-settings==2.3.4

# 密码哈希
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# JWT Token
python-jose[cryptography]==3.3.0

# 支持表单数据
python-multipart==0.0.9

# 数据库驱动
aiosqlite==0.20.0

# 其他依赖
python-dotenv==1.0.1